# api-gateway/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: |-
    Gateway for Mallow algorithmic trading system.

    All protected routes require a Bearer JWT token obtained from the Identity service (`/api/v1/auth/login`).
  title: Mallow API Gateway
  version: "1.0"
paths:
...
  /api/v1/hands:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List hands
      tags:
      - Helm
    post:
      consumes:
      - application/json
      description: Creates a new signal-following bot under a helm. The hand starts
        stopped; call /start to activate.
      parameters:
      - description: Hand config
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a hand (autonomous bot)
      tags:
      - Helm
  /api/v1/hands/{id}:
    delete:
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a hand
      tags:
      - Helm
    get:
      parameters:
      - description: Hand ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a hand
      tags:
      - Helm
    put:
      consumes:
      - application/json
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated config
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update hand config
      tags:
      - Helm
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"